cerebro/requirements.txt
setuptools >= 38.5.2
Django==1.11.0
Fabric==1.8.1 #note this needs to be fabric 2
reportlab==3.4.0

coverage==3.7
django-axes
django-braces
django-compressor
django-crispy-forms
django-datatables-view
django-debug-toolbar
django-extensions
django-grappelli
django-sticky-messages
django-storages
django-taggit
django-timezone-field
django-wysiwyg-redactor
Fabric==1.8.1
factory-boy==2.2.1
gunicorn==18.0
ipdb==0.8
ipython==1.1.0
mailchimp==2.0.9
newrelic==2.22.1.20
psycopg2==2.7.4
pycrypto==2.6.1
pytz==2013.7
reportlab==3.4.0
requests==2.3.0

# additional
django-model-utils
django-tables2
djangorestframework

# for heroku static files
whitenoise

# for production
dj-database-url
